code-generator: rename _test to _examples
