+++ staging/src/k8s.io/code-generator/_examples/apis/testgroup/types.go
