+++ staging/src/k8s.io/code-generator/_examples/apis/testgroup/v1/types.go
