+++ staging/src/k8s.io/code-generator/_examples/clientset/internal/fake/clientset_generated.go
