+++ staging/src/k8s.io/code-generator/_examples/clientset/internal/scheme/register.go
