+++ staging/src/k8s.io/code-generator/_examples/clientset/internal/typed/testgroup/internalversion/fake/doc.go
