+++ staging/src/k8s.io/code-generator/_examples/clientset/versioned/clientset.go
