+++ staging/src/k8s.io/code-generator/_examples/clientset/versioned/fake/clientset_generated.go
