+++ staging/src/k8s.io/code-generator/_examples/clientset/versioned/typed/testgroup/v1/fake/doc.go
