+++ staging/src/k8s.io/code-generator/_examples/informers/externalversions/generic.go
