+++ staging/src/k8s.io/code-generator/_examples/informers/internalversion/generic.go
