+++ staging/src/k8s.io/code-generator/_examples/informers/internalversion/internalinterfaces/factory_interfaces.go
