+++ staging/src/k8s.io/code-generator/_examples/informers/internalversion/testgroup/interface.go
