+++ staging/src/k8s.io/code-generator/_examples/listers/testgroup/internalversion/expansion_generated.go
