+++ staging/src/k8s.io/code-generator/_examples/listers/testgroup/internalversion/testtype.go
