+++ staging/src/k8s.io/code-generator/_examples/listers/testgroup/v1/testtype.go
